2020.05.31 - Many Many changes
Upon a macOS update with pc reboot during night, I don't now what's happened but all the source written and successfully compiled last night was unable to compile.
Arduino had broken indexes and didn't recognize some already installed libraries.
So I had to mend up all library folder and reinstall each one.
In this operations something was changed because some pieces of code got error on compiling as the methos I've called wasn't implemented in library...
It was so weird!
I clean the code and step-by-step get it compiling successfully.
I had a problem with DS3231 library when setting year in RTC clock... it convert number from DEC to BCD and wrote to RTC but, for some reason, year was written ad 2068 instead of 2020...
So I tried another DS3231 library... the DS3231M because I noticed that the name of my RTC chip was just DS3231M.
But DS3231 and DS3231M can't co-exist together, so I've had to comment out all code of the old one. Removing it is scheduled in the future.